gn: change properties of save() and load() methods
The method save() will not have a "targets" attribute now. That was
supposed to be used in order save a graph to be reused to execute the
nodes referred by it. We now have mingraph() for that.

The method load() will be a static method that, in turns, instantiate
Graph correctly.